fix(GiniHealthSDK): Remove test logs, remove unused methods, code cleanup
IPC-681
